requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
beautifulsoup4==4.12.2
reportlab==4.0.4
python-dotenv==1.0.0
gunicorn==21.2.0
